// Source/Types.ts
// -------------------------------------------------------------------------- //
//                               *       +                                    //
//                         '                  |                               //
//                     ()    .-.,="``"=.    - o -                             //
//                           '=/_       \\     |                              //
//                        *   |  '=._    |                                    //
//                             \\     `=./`,        '                         //
//                          .   '=.__.=' `='      *                           //
//                                                                            //
//                                                                            //
// File      : Types.ts                                                       //
// Project   : Saturno.FancyLib                                              //
// Date      : 2026-05-19                                                     //
// Copyright : Saturno Software - 2026                                        //
// Author    : mateusdigital <hello@mateus.digital>                           //
// -------------------------------------------------------------------------- //
// SPDX-License-Identifier: GPL-3.0-only

// -----------------------------------------------------------------------------
export interface CommentSyntax {
  singleLineStart: string;
  singleLineEnd: string;
  multiLineStart: string;
  multiLineMiddle: string;
  multiLineEnd: string;
}